requirements.txt
streamlit
mysql-connector-python
requests
pandas
plotly
